+++ src/Modules/Tours/Explorer.Tours.Core/Domain/Tours/Tour.cs
@@ -210,7 +210,7 @@ public class Tour : Entity
         return Status != Status.Published;
     }
 
-    internal bool IsUserAuthor(int userId)
+    internal bool IsUserAuthor(long userId)
     {
         return AuthorId==userId;
     }
